@@ -12,82 +12,53 @@ import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
-import android.widget.Button;
 import android.widget.ListView;
 
-import com.inappropirates.lightwalker.bluetooth.BluetoothStatusEnum;
-import com.inappropirates.lightwalker.bluetooth.SendSettingsThread;
-import com.inappropirates.lightwalker.config.Config;
-import com.inappropirates.lightwalker.config.Preferences;
-import com.inappropirates.lightwalker.bluetooth.BluetoothHandler;
+import com.inappropirates.lightwalker.bluetooth.BluetoothUartManager;
+import com.inappropirates.lightwalker.config.ModeManager;
 import com.inappropirates.lightwalker.ui.ModeListAdapter;
-import com.inappropirates.lightwalker.bluetooth.BluetoothBoss;
 import com.inappropirates.lightwalker.util.Util;
 
 public class MainActivity extends AppCompatActivity {
-    private static BluetoothBoss btBoss;
-    private static Context context;
-    private ListView modeListView;
-    private BluetoothHandler bluetoothHandler;
+    private Context context;
     private BroadcastReceiver bluetoothDisconnectReceiver;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+
+        context = this;
         setContentView(R.layout.activity_main);
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
 
-        final Button bluetoothButton = (Button) findViewById(R.id.bluetooth_button);
-        bluetoothButton.setOnClickListener(new View.OnClickListener()
-        {
-            @Override
-            public void onClick(View view)
-            {
-                bluetoothButton.setText("connecting");
-                bluetoothButton.setBackgroundColor(android.graphics.Color.argb(255, 246, 163, 85));
-                bluetoothButton.setEnabled(false);
+        BluetoothUartManager.INSTANCE.connect();
 
-                if (btBoss == null)
-                {
-                    btBoss = new BluetoothBoss(context, bluetoothHandler);
-                    btBoss.connect();
-                }
-            }
-        });
-
-        modeListView = (ListView) findViewById(R.id.modeListView);
-        modeListView.setAdapter(new ModeListAdapter(this, Config.modes));
+        ListView modeListView = findViewById(R.id.modeListView);
+        modeListView.setAdapter(new ModeListAdapter(this, ModeManager.INSTANCE.getModes()));
     }
 
     @Override
     protected void onStart() {
         super.onStart();
-
-        context = this;
-        bluetoothHandler = new BluetoothHandler((Button) findViewById(R.id.bluetooth_button), context);
-
-        if (btBoss != null)
-        {
-            btBoss.setHandler(bluetoothHandler);
-            bluetoothHandler.setConnectedButton(btBoss.getState() == BluetoothStatusEnum.CONNECTED);
-        }
+        BluetoothStatusHandler bluetoothStatusHandler = new BluetoothStatusHandler(findViewById(R.id.bluetooth_button), context);
+        BluetoothUartManager.INSTANCE.setHandler(bluetoothStatusHandler);
+        BluetoothUartManager.INSTANCE.updateStatus();
     }
 
     @Override
     protected void onResume() {
         super.onResume();
 
-        bluetoothDisconnectReceiver = new BroadcastReceiver()
-        {
+        BluetoothUartManager.INSTANCE.updateStatus();
+
+        bluetoothDisconnectReceiver = new BroadcastReceiver() {
             @Override
-            public void onReceive(Context context, Intent intent)
-            {
+            public void onReceive(Context context, Intent intent) {
                 String action = intent.getAction();
-                if (action.equals(BluetoothDevice.ACTION_ACL_DISCONNECTED))
-                {
-                    if (btBoss != null)
-                        btBoss.stop();
+                assert action != null;
+                if (action.equals(BluetoothDevice.ACTION_ACL_DISCONNECTED)) {
+                    BluetoothUartManager.INSTANCE.updateStatus();
                 }
             }
         };
@@ -95,8 +66,7 @@ public class MainActivity extends AppCompatActivity {
     }
 
     @Override
-    protected void onPause()
-    {
+    protected void onPause() {
         super.onPause();
         context.unregisterReceiver(bluetoothDisconnectReceiver);
     }
@@ -105,10 +75,8 @@ public class MainActivity extends AppCompatActivity {
     protected void onDestroy() {
         super.onDestroy();
 
-        if (isFinishing() && btBoss != null) {
-            btBoss.stop();
-            btBoss = null;
-        }
+        if (isFinishing())
+            BluetoothUartManager.INSTANCE.disconnect();
     }
 
     @Override
@@ -122,8 +90,7 @@ public class MainActivity extends AppCompatActivity {
     public boolean onOptionsItemSelected(MenuItem item) {
         int id = item.getItemId();
 
-        switch (id)
-        {
+        switch (id) {
             case R.id.action_settings:
                 Log.d(Util.TAG, "settings!");
                 Intent intent = new Intent(context, SettingsActivity.class);
@@ -132,37 +99,17 @@ public class MainActivity extends AppCompatActivity {
                 return true;
             case R.id.action_disconnect:
                 Log.d(Util.TAG, "disconnect!");
-                if (btBoss != null)
-                {
-                    btBoss.stop();
-                    btBoss = null;
-                }
+                BluetoothUartManager.INSTANCE.disconnect();
                 return true;
         }
 
         return super.onOptionsItemSelected(item);
     }
 
-    public static void sendSetting(String key, String value)
-    {
-        if (btBoss != null)
-        {
-            Preferences preference = Preferences.valueOf(key);
-            String msg = String.format("sending %s(%d)=%s... ", key, preference.ordinal(), value);
-            if (btBoss.send(String.format("%d=%s\r", preference.ordinal(), value)))
-                Log.d(Util.TAG, msg + "success!");
-            else
-                Log.d(Util.TAG, msg + "failed.");
-        }
+    public void connectButtonClicked(View view) {
+        view.setEnabled(false);
+        BluetoothUartManager.INSTANCE.connect();
     }
 
-    public static void sendAllCurrentSettings()
-    {
-        if (btBoss != null)
-        {
-            SendSettingsThread sendSettingsThread = new SendSettingsThread(Config.currentMode, context);
-            sendSettingsThread.start();
-        }
-    }
 
 }
