@@ -4,14 +4,13 @@ import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.preference.PreferenceFragment;
 
-import com.inappropirates.lightwalker.MainActivity;
+import com.inappropirates.lightwalker.bluetooth.BluetoothUartManager;
 import com.inappropirates.lightwalker.config.Mode;
-import com.inappropirates.util.PropertyFormatter;
+import com.inappropirates.lightwalker.util.PropertyFormatter;
 
 // <cgerstle> PreferenceFragment wants you to have a default constructor and a setter instead of
 // a non-default constructor... not really sure that applies since I'm creating manually, but whatev
-public class SettingsFragment extends PreferenceFragment implements SharedPreferences.OnSharedPreferenceChangeListener
-{
+public class SettingsFragment extends PreferenceFragment implements SharedPreferences.OnSharedPreferenceChangeListener {
     private Mode mode;
 
     public Mode getMode() {
@@ -30,23 +29,20 @@ public class SettingsFragment extends PreferenceFragment implements SharedPrefer
     }
 
     @Override
-    public void onResume()
-    {
+    public void onResume() {
         super.onResume();
         getPreferenceManager().getSharedPreferences().registerOnSharedPreferenceChangeListener(this);
     }
 
     @Override
-    public void onPause()
-    {
+    public void onPause() {
         super.onPause();
         getPreferenceManager().getSharedPreferences().unregisterOnSharedPreferenceChangeListener(this);
     }
 
     @Override
-    public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key)
-    {
+    public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key) {
         Object value = sharedPreferences.getAll().get(key);
-        MainActivity.sendSetting(key, PropertyFormatter.getStringVal(key, value));
+        BluetoothUartManager.INSTANCE.sendSetting(key, PropertyFormatter.getStringVal(key, value));
     }
 }
